Polish "Detect Persistent and RelationshipProperties with Neo4j"
See gh-24239
