+++ spring-boot-project/spring-boot-autoconfigure/src/test/java/org/springframework/boot/autoconfigure/data/neo4j/Neo4jReactiveDataAutoConfigurationTests.java
@@ -21,6 +21,11 @@ import reactor.core.publisher.Mono;
 import reactor.test.StepVerifier;
 
 import org.springframework.boot.autoconfigure.AutoConfigurations;
+import org.springframework.boot.autoconfigure.TestAutoConfigurationPackage;
+import org.springframework.boot.autoconfigure.data.neo4j.scan.TestNode;
+import org.springframework.boot.autoconfigure.data.neo4j.scan.TestNonAnnotated;
+import org.springframework.boot.autoconfigure.data.neo4j.scan.TestPersistent;
+import org.springframework.boot.autoconfigure.data.neo4j.scan.TestRelationshipProperties;
 import org.springframework.boot.autoconfigure.neo4j.Neo4jAutoConfiguration;
 import org.springframework.boot.test.context.runner.ApplicationContextRunner;
 import org.springframework.context.annotation.Bean;
@@ -30,6 +35,7 @@ import org.springframework.data.neo4j.core.ReactiveDatabaseSelectionProvider;
 import org.springframework.data.neo4j.core.ReactiveNeo4jClient;
 import org.springframework.data.neo4j.core.ReactiveNeo4jOperations;
 import org.springframework.data.neo4j.core.ReactiveNeo4jTemplate;
+import org.springframework.data.neo4j.core.mapping.Neo4jMappingContext;
 import org.springframework.transaction.ReactiveTransactionManager;
 import org.springframework.transaction.TransactionManager;
 
@@ -121,6 +127,17 @@ class Neo4jReactiveDataAutoConfigurationTests {
 						.hasSingleBean(TransactionManager.class));
 	}
 
+	@Test
+	void shouldFilterInitialEntityScanWithKnownAnnotations() {
+		this.contextRunner.withUserConfiguration(EntityScanConfig.class).run((context) -> {
+			Neo4jMappingContext mappingContext = context.getBean(Neo4jMappingContext.class);
+			assertThat(mappingContext.hasPersistentEntityFor(TestNode.class)).isTrue();
+			assertThat(mappingContext.hasPersistentEntityFor(TestPersistent.class)).isTrue();
+			assertThat(mappingContext.hasPersistentEntityFor(TestRelationshipProperties.class)).isTrue();
+			assertThat(mappingContext.hasPersistentEntityFor(TestNonAnnotated.class)).isFalse();
+		});
+	}
+
 	@Configuration(proxyBeanMethods = false)
 	static class CustomReactiveDatabaseSelectionProviderConfiguration {
 
@@ -128,6 +145,12 @@ class Neo4jReactiveDataAutoConfigurationTests {
 		ReactiveDatabaseSelectionProvider databaseNameProvider() {
 			return () -> Mono.just(DatabaseSelection.byName("custom"));
 		}
+
+	}
+
+	@Configuration(proxyBeanMethods = false)
+	@TestAutoConfigurationPackage(TestPersistent.class)
+	static class EntityScanConfig {
 
 	}
