+++ spring-boot-project/spring-boot-autoconfigure/src/test/java/org/springframework/boot/autoconfigure/data/neo4j/scan/TestNode.java
@@ -16,12 +16,15 @@
 
 package org.springframework.boot.autoconfigure.data.neo4j.scan;
 
-import org.springframework.data.neo4j.core.schema.RelationshipProperties;
+import org.springframework.data.neo4j.core.schema.GeneratedValue;
+import org.springframework.data.neo4j.core.schema.Id;
+import org.springframework.data.neo4j.core.schema.Node;
 
-/**
- * @author Gerrit Meier
- */
-@RelationshipProperties
-public class AnnotatedWithRelationshipProperties {
+@Node
+public class TestNode {
+
+	@Id
+	@GeneratedValue
+	private Long id;
 
 }
